Updated constructor for Ball class. Added movement and collision detection for ball.

+++ include/game_constants.h
@@ -8,7 +8,7 @@ namespace game_constants
 
     constexpr unsigned int x_distance_from_border_to_paddle = 100;
 
-    const double dt = 0.01;
+    const float dt = 0.01;
 }
 
 #endif // GAME_CONSTANTS_H_INCLUDED
